fix batch_size dtype

+++ elk_generalization/training/run_sft.py
@@ -24,25 +24,25 @@ models = [
 ]
 
 ds_names = [
-    ("capitals", 4.0, 1.0),
-    ("hemisphere", 1.0, 1.0),
-    ("population", 2.0, 1.0),
+    ("capitals", 4.0, 1),
+    ("hemisphere", 1.0, 1),
+    ("population", 2.0, 1),
     ("sciq", 2.0, 1 / 16),
     ("sentiment", 2.0, 1 / 8),
     ("nli", 4.0, 1 / 8),
     ("authors", 4.0, 1 / 2),
-    ("addition", 1.0, 1.0),
-    ("subtraction", 1.0, 1.0),
-    ("multiplication", 1.0, 1.0),
-    ("modularaddition", 2.0, 1.0),
-    ("squaring", 1.0, 1.0),
+    ("addition", 1.0, 1),
+    ("subtraction", 1.0, 1),
+    ("multiplication", 1.0, 1),
+    ("modularaddition", 2.0, 1),
+    ("squaring", 1.0, 1),
 ]
 
 model, epoch_multiplier1, bs_multiplier1 = models[rank // len(ds_names)]
 ds_name, epoch_multiplier2, bs_multiplier2 = ds_names[rank % len(ds_names)]
 
 num_epochs = 15 * epoch_multiplier1 * epoch_multiplier2
-batch_size = max(bs_multiplier1 * bs_multiplier2, 1.0)
+batch_size = int(max(bs_multiplier1 * bs_multiplier2, 1))
 accum_steps = 32 // batch_size
 
 model_last = model.split("/")[-1]
